Save refactoring
- added external class to save mushroom format
- some code cleanup, preparing for full_save option
