+++ mushroom_rl/core/__init__.py
@@ -1,3 +1,4 @@
 from .core import Core
+from .serialization import Serializable
 
-__all__ = ['Core']
+__all__ = ['Core', 'Serializable']
